requirements.txt
anthropic
requests
edge-tts
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
